Improve sysconfig/network logic on RHEL
* This changes the logic around managing /etc/sysconfig/network
on RHEL. The file will be managed only if one of its related params
are set. Using the gateway param as a trigger is no longer required.
* It is most common to use FQDN as the HOSTNAME value. Fixes #35
* This change is not 100% backwards compatible. Up until this commit,
if you set gateway you'd get a default value for HOSTNAME too. After
this commit, the default no longer applies and will be removed.