ensure bond_mode and bond_miimon are respected on Redhat
If $bonding_opts is not set, it makes sense to use the values for $bond_mode
and $bond_miimon on Redhat, as these are the two options that are typically
set in $bonding_opts. This would also align with what a user expects when
setting $bond_mode and $bond_miimon.