Bugfix: remove duplicate defintion for Exec[network_restart_${name}]
if network::interface and network::mroute is used for the same
interface, and $restart_all_nics is false, the definition for
network_restart will be duplicated, using puppet to fail.

if you use
  network::mroute { 'static-eth0':
    interface => 'eth0'
    ...
  }
puppet will not complain, but create a "route-static-eth0"-file which
will not be pickuped by networkmanager.

So either move the exec-definition into a separate file (and use it from
with the interface.pp and mroute.pp manifest), or check if the
exec-definition was already defined, or (the way I choosed for now) use
a different name for the exec-ressource.